requirements.txt
torch==2.0.1+cu118  # GPU-optimized PyTorch
transformers>=4.30.0
huggingface_hub>=0.16.4
pinecone-client>=2.2.1
opencv-python>=4.8.0
unrealcv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
Pillow>=10.0.0
grpcio>=1.56.0
protobuf>=4.23.0
palm-e>=2.0.0
whisper-v3>=1.0.0
fairscale
fire
tiktoken==0.4.0
blobfile
bitsandbytes>=0.37.0
fairscale>=0.4.0
langchain>=0.0.200
